remove some tables and models to make the project simple

+++ database/migrations/2025_04_27_193507_create_served_customers_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('queue_id')->nullable()->constrained()->onDelete('set null');
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->integer('waiting_time');
+            $table->float('waiting_time');
             $table->timestamps();
         });
     }
